==> legacy_code/01_create_and_save_light_schedules.ipynb <==
Since `Chang14` light schedules with small `dt` take long to generate, we calculate them once and save the results. This notebook generates the time grids and saves them in `data/light_schedules/`.

import os
import numpy as np
from circadian.lights import LightSchedule

save_path = "data/light_schedules"
if not os.path.exists(save_path):
    os.makedirs(save_path)

dt = 0.005 # hours
days = 14.0
# Regular schedule parameters early birds
regular_indoor_lux = 1000.0
regular_parameters_early = {
    "lights_on": 6.0,
    "lights_off": 22.0,
}
# Regular schedule parameters night owls
regular_parameters_late = {
    "lights_on": 8.0,
    "lights_off": 24.0,
}
# Chang et al. 2014 conditions
chang14_parameters = {
    "typical_indoor_lux": 90.0,
    "dim_lux": 3.0,
    "ereader_lux": 31.73,
    "book_lux": 0.91,
    "reading_start_time": 18.0,
    "reading_duration": 4.0,
}
# Chang et al. 2014 + high light conditions
chang14_high_parameters = {
    "typical_indoor_lux": 500.0,
    "dim_lux": 3.0,
    "ereader_lux": 31.73,
    "book_lux": 0.91,
    "reading_start_time": 18.0,
    "reading_duration": 4.0,
}
# Realistic conditions
realistic_parameters = {
    "typical_indoor_lux": 500.0, # change here
    "dim_lux": 3.0,
    "ereader_lux": 31.73,
    "book_lux": 0.91,
    "reading_start_time": 21.0, # change here
    "reading_duration": 1.0, # change here
}

time = np.arange(0, 24*days, dt)

regular_early_schedule = LightSchedule.Regular(regular_indoor_lux, **regular_parameters_early)
regular_late_schedule = LightSchedule.Regular(regular_indoor_lux, **regular_parameters_late)

chang14_ebook_first_schedule = LightSchedule.Chang14(**chang14_parameters,
                                                     first_reading_condition="eReader")
chang14_ebook_second_schedule = LightSchedule.Chang14(**chang14_parameters,
                                                      first_reading_condition="Book")
chang14_high_ebook_first_schedule = LightSchedule.Chang14(**chang14_high_parameters,
                                                          first_reading_condition="eReader")
chang14_high_ebook_second_schedule = LightSchedule.Chang14(**chang14_high_parameters,
                                                           first_reading_condition="Book")
realistic_ebook_first_schedule = LightSchedule.Chang14(**realistic_parameters,
                                                       first_reading_condition="eReader")
realistic_ebook_second_schedule = LightSchedule.Chang14(**realistic_parameters,
                                                        first_reading_condition="Book")

regular_early_light = regular_early_schedule(time)
regular_late_light = regular_late_schedule(time)

chang14_ebook_first_light = chang14_ebook_first_schedule(time)
chang14_ebook_second_light = chang14_ebook_second_schedule(time)

chang14_high_first_light = chang14_high_ebook_first_schedule(time)
chang14_high_second_light = chang14_high_ebook_second_schedule(time)

realistic_ebook_first_light = realistic_ebook_first_schedule(time)
realistic_ebook_second_light = realistic_ebook_second_schedule(time)

np.savez(f"{save_path}/parameters_light_schedules.npz",
         regular_indoor_lux=regular_indoor_lux,
         regular_early_parameters=regular_parameters_early,
         regular_late_parameters=regular_parameters_late,
         chang14_parameters=chang14_parameters,
         chang14_high_parameters=chang14_high_parameters,
         realistic_parameters=realistic_parameters)

np.save(f"{save_path}/time.npy", time)
np.save(f"{save_path}/regular_early_light.npy", regular_early_light)
np.save(f"{save_path}/regular_late_light.npy", regular_late_light)
np.save(f"{save_path}/chang14_ebook_first_light.npy", chang14_ebook_first_light)
np.save(f"{save_path}/chang14_ebook_second_light.npy", chang14_ebook_second_light)
np.save(f"{save_path}/chang14_high_ebook_first_light.npy", chang14_high_first_light)
np.save(f"{save_path}/chang14_high_ebook_second_light.npy", chang14_high_second_light)
np.save(f"{save_path}/realistic_ebook_first_light.npy", realistic_ebook_first_light)
np.save(f"{save_path}/realistic_ebook_second_light.npy", realistic_ebook_second_light)
