==> legacy_code/07a_process_selected_parameters_simulations.ipynb <==
import os
import numpy as np
import pandas as pd
from utils.helper_functions import calculate_cbtmin, calculate_sleep_onset, calculate_sleep_duration

save_path = f"data/chang14_selected_sleep_parameters"
if not os.path.exists(save_path):
    os.makedirs(save_path)

selected_parameters_early = pd.read_csv('data/sleep_parameter_exploration/early_light_selected_sleep_parameters.csv')
selected_parameters_late = pd.read_csv('data/sleep_parameter_exploration/late_light_selected_sleep_parameters.csv')
selected_parameters_options = [selected_parameters_early, selected_parameters_late]

time = np.load("data/light_schedules/time.npy")

DAILY_CONDITION_EBOOK_FIRST = [
    'None',
    'eBook', 'eBook', 'eBook', 'eBook', 'eBook',
    'None',
    'Book', 'Book', 'Book', 'Book', 'Book',
    'None',
]
DAILY_CONDITION_EBOOK_SECOND = [
    'None',
    'Book', 'Book', 'Book', 'Book', 'Book',
    'None',
    'eBook', 'eBook', 'eBook', 'eBook', 'eBook',
    'None',
]

DAILY_CONDITION_CHOICE = {
    'regular': ['None'] * 13,
    'chang14_ebook_first': DAILY_CONDITION_EBOOK_FIRST,
    'chang14_ebook_second': DAILY_CONDITION_EBOOK_SECOND,
    'chang14_high_ebook_first': DAILY_CONDITION_EBOOK_FIRST,
    'chang14_high_ebook_second': DAILY_CONDITION_EBOOK_SECOND,
    'realistic_ebook_first': DAILY_CONDITION_EBOOK_FIRST,
    'realistic_ebook_second': DAILY_CONDITION_EBOOK_SECOND,
}

light_conditions = [
    'early_light', 'late_light'
]

all_reading_conditions = [
    'chang14_ebook_first', 'chang14_ebook_second',
    'chang14_high_ebook_first', 'chang14_high_ebook_second',
    'realistic_ebook_first', 'realistic_ebook_second',
]
simulation_data_path = "data/chang14_selected_sleep_parameters"

simulation_features = pd.DataFrame()
for jdx, light_condition in enumerate(light_conditions):
    selected_parameters = selected_parameters_options[jdx]
    for idx in selected_parameters.index:
        data = np.load(f"{simulation_data_path}/{light_condition}/{light_condition}_chang14_simulation_{idx}.npz", 
                        allow_pickle=True)

        regular_sleep_onset = calculate_sleep_onset(time, data, 'regular')
        regular_cbtmin = calculate_cbtmin(time, data, 'regular')
        regular_sleep_duration = calculate_sleep_duration(time, data, 'regular')

        # Check if they have the same size
        min_length = min(len(regular_sleep_onset), len(regular_cbtmin))
        regular_sleep_onset = regular_sleep_onset[:min_length]
        regular_cbtmin = regular_cbtmin[:min_length]
        regular_sleep_duration = regular_sleep_duration[:min_length]

        normalized_sleep_onset = regular_sleep_onset - np.arange(1, len(regular_sleep_onset) + 1) * 24 + 24
        normalized_cbtmin = regular_cbtmin - np.arange(1, len(regular_cbtmin) + 1) * 24 + 24
        normalized_sleep_latency = (normalized_sleep_onset - 22.0) * 60
        normalized_sleep_latency[normalized_sleep_latency < 0] = 0.0

        daily_condition = DAILY_CONDITION_CHOICE['regular']

        data_subset = {
            'simulation_idx': idx,
            'day': np.arange(1, len(regular_cbtmin) + 1),
            'simulation_condition': 'regular',
            'sleep_onset': normalized_sleep_onset,
            'sleep_duration': regular_sleep_duration,
            'cbtmin': normalized_cbtmin,
            'sleep_latency_minutes': normalized_sleep_latency,
            'daily_condition': daily_condition[:min_length],
            'light_condition': light_condition,
        }

        simulation_features = pd.concat([simulation_features, pd.DataFrame(data_subset)])

        for reading_condition in all_reading_conditions:
            sleep_onset = calculate_sleep_onset(time, data, reading_condition)
            cbtmin = calculate_cbtmin(time, data, reading_condition)
            sleep_duration = calculate_sleep_duration(time, data, reading_condition,
                                                      light_condition=light_condition)

            min_length = min(len(sleep_onset), len(cbtmin))
            sleep_onset = sleep_onset[:min_length]
            cbtmin = cbtmin[:min_length]
            sleep_duration = sleep_duration[:min_length]

            normalized_sleep_onset = sleep_onset - np.arange(1, len(sleep_onset) + 1) * 24 + 24
            normalized_cbtmin = cbtmin - np.arange(1, len(cbtmin) + 1) * 24 + 24
            normalized_sleep_latency = (normalized_sleep_onset - 22.0) * 60
            normalized_sleep_latency = (sleep_onset % 24 - 22.0) * 60
            normalized_sleep_latency[normalized_sleep_latency < 0] = 0.0

            daily_condition = DAILY_CONDITION_CHOICE[reading_condition]

            data_subset = {
                'simulation_idx': idx,
                'day': np.arange(1, len(cbtmin) + 1),
                'simulation_condition': reading_condition,
                'sleep_onset': normalized_sleep_onset,
                'sleep_duration': sleep_duration,
                'cbtmin': normalized_cbtmin,
                'sleep_latency_minutes': normalized_sleep_latency,
                'daily_condition': daily_condition[:min_length],
                'light_condition': light_condition,
            }
            simulation_features = pd.concat([simulation_features, pd.DataFrame(data_subset)])
    
simulation_features.to_csv(f"{save_path}/selected_sleep_parameters_simulation_features.csv",
                           index=False)
